+++ Foreman/Models/Nodes/SupplierNode.cs
@@ -23,8 +23,8 @@ namespace Foreman
 			SuppliedItem = item;
 		}
 
-		public override float GetConsumeRate(Item item) { throw new ArgumentException("Supplier does not consume! nothing should be asking for the consume rate"); }
-		public override float GetSupplyRate(Item item) { return (float)Math.Round(ActualRate, RoundingDP); }
+		public override double GetConsumeRate(Item item) { throw new ArgumentException("Supplier does not consume! nothing should be asking for the consume rate"); }
+		public override double GetSupplyRate(Item item) { return (double)Math.Round(ActualRate, RoundingDP); }
 
 		internal override double inputRateFor(Item item) { throw new ArgumentException("Supplier should not have outputs!"); }
 		internal override double outputRateFor(Item item) { return 1; }
